+++ Components/MeshLodGenerator/src/OgreLodCollapser.cpp
@@ -117,7 +117,7 @@ namespace Ogre
     }
 #endif // if OGRE_DEBUG_MODE
 
-    bool LodCollapser::hasSrcID(unsigned int srcID, unsigned short submeshID)
+    bool LodCollapser::hasSrcID(unsigned int srcID, unsigned submeshID)
     {
         // This will only return exact matches.
         for (size_t i = 0; i < tmpCollapsedEdges.size(); i++)
@@ -153,7 +153,7 @@ namespace Ogre
         }
     }
 
-    size_t LodCollapser::findDstID(unsigned int srcID, unsigned short submeshID)
+    size_t LodCollapser::findDstID(unsigned int srcID, unsigned submeshID)
     {
         // Tries to find a compatible edge.
